feat: async smtp and verification resend (#101)
* feat: enable async for smtp

* feat: add TooManyRequestsException and handler

* feat: add tooManyRequests to EmailVerificationExceptionFactory

* feat: add getToken in VerificationMailMessage

* feat: add cooldown and reuse valid verification tokens

* feat: add ResendVerification domain model

* feat: add ResendVerificationResult domain model

* feat: add validation for ResendVerification model

* feat: implement ResendVerificationUseCase

* feat: add ResendVerificationRequest and Response

* feat: add mappers to AuthWebMapper

* feat: add resend-code endpoint to AuthController

+++ src/main/java/ru/jerael/booktracker/backend/application/service/EmailVerificationServiceImpl.java
@@ -3,15 +3,17 @@ package ru.jerael.booktracker.backend.application.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.jerael.booktracker.backend.application.validator.VerificationTokenValidatorImpl;
 import ru.jerael.booktracker.backend.domain.exception.factory.EmailVerificationExceptionFactory;
 import ru.jerael.booktracker.backend.domain.mail.VerificationMailMessage;
 import ru.jerael.booktracker.backend.domain.model.verification.*;
 import ru.jerael.booktracker.backend.domain.repository.EmailVerificationRepository;
 import ru.jerael.booktracker.backend.domain.smtp.SmtpService;
+import ru.jerael.booktracker.backend.domain.validator.VerificationTokenValidator;
 import ru.jerael.booktracker.backend.domain.verification.EmailVerificationService;
 import ru.jerael.booktracker.backend.domain.verification.VerificationTokenGenerator;
+import java.time.Duration;
 import java.time.Instant;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -20,7 +22,7 @@ public class EmailVerificationServiceImpl implements EmailVerificationService {
     private final VerificationTokenGenerator verificationTokenGenerator;
     private final EmailVerificationRepository emailVerificationRepository;
     private final SmtpService smtpService;
-    private final VerificationTokenValidatorImpl verificationTokenValidator;
+    private final VerificationTokenValidator verificationTokenValidator;
 
     @Override
     @Transactional
@@ -29,7 +31,20 @@ public class EmailVerificationServiceImpl implements EmailVerificationService {
         String email = payload.email();
         VerificationType type = payload.type();
 
-        emailVerificationRepository.deleteByUserIdAndType(userId, type);
+        Optional<EmailVerification> found = emailVerificationRepository.findByUserIdAndType(userId, type);
+        if (found.isPresent()) {
+            EmailVerification verification = found.get();
+            if (Instant.now().isBefore(verification.createdAt().plusSeconds(60))) {
+                throw EmailVerificationExceptionFactory.tooManyRequests();
+            }
+            if (Instant.now().isBefore(verification.expiresAt())) {
+                Duration remain = Duration.between(Instant.now(), verification.expiresAt());
+                VerificationToken token = new VerificationToken(verification.token(), remain);
+                smtpService.sendEmail(email, new VerificationMailMessage(token));
+                return verification.expiresAt();
+            }
+            emailVerificationRepository.deleteByUserIdAndType(userId, type);
+        }
 
         VerificationToken token = verificationTokenGenerator.generate();
